+++ CHANGELOG.md
@@ -11,8 +11,7 @@
 ## v0.2
 
 ### v0.2.32
-- 🔶 deprecate `attempt` in favor of `resultify`
-- 🔶 deprecate `attemptAsync` in favor of `resultifyAsync`
+- 🍏 add `resultify` and `resultifyAsync` fns
 
 ### v0.2.31
 - 🔶 make err agnostic about error types